New conversion script: nodejs.sh
Building Chromium on jammy now requires a newer Node.js interpreter
than what Ubuntu provides, and backporting a newer version of "nodejs"
is non-trivial. This script is intended to apply not just to that
package, but also the following build dependencies: pkg-js-tools,
node-undici, node-cjs-module-lexer.

Source for the following packages should be pulled from lunar or
later, and built in this order: node-busboy, pkg-js-tools,
node-undici, node-cjs-module-lexer, nodejs.

Some associated updates in functions.sh: Fix the multiple
XSBC-Original-Maintainer: field check to account for debian.net e-mail
addresses; use a variable so that the script can modify the changelog
text if desired; account for packages that do not have a patch series
(e.g. pkg-js-tools)