firefox-focal: Clean up Clang/LLVM build-deps additions
The upstream package can build (as of this writing) with Clang/LLVM
versions 13, 12, or 10 (preferenced in that order). 13 is available
in jammy up to mantic; the lowest available in noble is 14. So we no
longer need to add 11+12+13+14 (as when the upstream package wanted
10), just 14.

Also, make the Firefox startup script's XDG_SESSION_TYPE check a bit
more robust.