chromium: Fix startup crash due to our hacked <optional> header
As of 124.0.6367.60, Chromium began crashing on startup due to this
assertion failure:

../../optional:439: _Tp &std::_Optional_base_impl<content::DocumentAssociatedData, std::_Optional_base<content::DocumentAssociatedData, false, false>>::_M_get() [_Tp = content::DocumentAssociatedData, _Dp = std::_Optional_base<content::DocumentAssociatedData, false, false>]: Assertion 'this->_M_is_engaged()' failed.

It's unclear why this is only happening now, but it suggests that the
approach taken in our xtradeb_absl-optional-libstdc++-11.patch (which
contains a modified full copy of the libstdc++-11 <optional> header) is
not sustainable.

Some further investigation of Debian's fixes/absl-optional.patch (which
our patch effectively duplicates, except Debian uses a newer version of
the <optional> header) found that a small tweak is enough to make it
usable on jammy. This renders our patch unnecessary.

xtradeb_absl-optional-fix.patch: New patch which tweaks Debian's hacked
<optional> header so that it compiles on jammy

xtradeb_absl-optional-libstdc++-11.patch: Drop, as maintaining a wholly
separate modified <optional> header is hard and unnecessary

chromium.sh: Update patch references accordingly