New conversion script: harfbuzz.sh
Chromium 124.0.6367.118 requires the harfbuzz-subset library, which is
not provided by the harfbuzz 2.7.4 package in Ubuntu jammy. We copied
in the sources for 6.0.0 from lunar to the deps/jammy repo, but the
shared-library dependency introduced by this build resulted in an
uninstallable package for users:

    https://bugs.launchpad.net/bugs/2064821

The libharfbuzz-dev package does not provide static libraries,
presumably because GObject introspection (which is enabled) requires
shared ones. This script thus modifies the package with the primary
intent of providing static libraries, plus other changes needed as a
consequence of that.

This script supports both harfbuzz 6.0.0 (in Debian stable) and 8.3.0
(unstable).