chromium: Fix libgtk dependency for jammy and mantic
The original dep is "libgtk-3-0t64 | xdg-desktop-portal-backend". As
the former package does not exist in jammy nor mantic, the effect is
to require installation of XDG portal infrastructure, which may not be
desired by the user. Edit the package name to "libgtk-3-0", to match
the dependency Debian uses on bookworm.