Revamp error handling
Define three new common functions, and update the conversion scripts
accordingly:

warning(): Print a warning message. Not used yet, but will be useful.

error(): Use this instead of variations on the "print error message
and exit 1" idiom. Simplify resulting program syntax where feasible.

not_applicable(): Special-purpose exit routine to indicate that the
Ubuntu release being targeted does not need the conversion, that the
conversion is not possible, or that for whatever other reason no
converted package should result. This will exit with a status of 2, to
indicate the different nature of the failure to convert.

Also, in functions.sh, move the Maintainer:/Uploaders: edits (in
$debian/control) from the initialize() to the finish() functions, so
that conversion scripts can bail out before any file modifications are
performed.