New utility: can-install.sh
(First item of a new util/ subdirectory, too!)

This script helps check whether packages in the XtraDeb repos can be
installed on their intended target release (e.g. "can all the packages
in the 'apps' repo for noble actually be installed on noble?"), or
potentially other/later target releases (e.g. "can all the packages in
the 'apps' repo for noble be installed on oracular?").

It's intended to help make Jhonny's life easier :)