chromium: Remove spurious libtest_trace_processor.so dependency
This warning turned up in the jammy build:

  dpkg-shlibdeps: warning: cannot find library
  libtest_trace_processor.so needed by
  debian/chromium-shell/usr/lib/chromium/chromium-shell (ELF format:
  'elf64-x86-64' abi: '0201003e00000000'; RPATH: '/usr/lib/chromium')

No other Ubuntu nor Debian release seems to run into this. The library
is always passed to the linker, but the resulting executable never
gets the dependency, except on jammy.

For lack of a cleaner way to avoid the dependency, hack the offending
.ninja file to physically remove the reference to the library.