rustc: Updates from production usage
Implement a better package-version check that takes into account the
differing versions of Rust supported in the various Ubuntu releases.
This helps to avoid the mistake of e.g. building 1.81 for oracular.

Also, remove $debian/source/options . The "include-removal" option
therein leads to the following error from our modified version of
dpkg-source(1) when we attempt to create a source package:

  dpkg-source-new: error: --no-generate-diff is incompatible with
  the following options: --include-removal, --include-timestamp,
  --include-binaries, --create-empty-orig, --auto-commit