firefox-focal is now firefox-mt
Ubuntu ended support for 20.04/focal back in April, and the non-snap
firefox package from this release was last updated in March to 136.0.
Thus a new source for this package is needed, and given that the
Mozilla Team PPA appears to have been the upstream provider of the
focal package (the debianization has minimal diffs between the two),
I've updated this conversion script to target their work.

A few additional changes along the way:

* Enable thin LTO in the build

* Drop the patching of the Wayland support bit into the startup
  script, as that has been merged upstream

* Allow unsigned extensions in system dirs (as Debian does)

* Reduce the frequency of keepalive messages during the build

* Drop the Rust version edits, as 1.80 is no longer used anyway (the
  Firefox build now requires >= 1.82)

* Drop the nodejs-mozilla edit, as this is no longer needed