firefox-mt: Use a single version of the LLVM build-dependencies
The Firefox package from Mozilla Team has these LLVM build-deps:

    clang-20 | clang-19 | clang-18,
    libclang-20-dev | libclang-19-dev | libclang-18-dev,
    llvm-20-dev | llvm-19-dev | llvm-18-dev,

This is, unfortunately, *too* flexible for our needs: it allows using
e.g. clang-19 with libclang-18-dev, which breaks the build. The CI
system that tests new packages prior to uploading them has a number of
build-deps pre-installed (to reduce turnaround time), and the above
approach does not play well with such a premise.

Thus, drop the alternations, and use a single LLVM version (the
specific version depending on the release being targeted).

Side note: What we sure could use is the ability to specify e.g.

    (clang-20 & libclang-20-dev & llvm-20-dev) |
    (clang-19 & libclang-19-dev & llvm-19-dev) |
    (clang-18 & libclang-18-dev & llvm-18-dev),

as that better captures the intent of the original build-dep lines.