Streamline foreign-arch container builds
Thanks to discussion on https://bugs.debian.org/1054104, I've learned
that it is not necessary to copy the QEMU user mode emulation binaries
into a foreign-arch container image in order to be able to run the
container. This simplifies the container build.

Also added a small check to ensure that "unstable" is used instead of
"sid" when specifying a Debian-based container.